Split FF heuristic into self-contained headers along its DAG
heuristic.hpp had two mid-file #include directives that reopened
namespace railroad three times, and heuristic_prob.hpp was a
non-self-contained fragment relying on its includer.

Split along the real dependency DAG into self-contained headers, each
with #pragma once and its own includes:

  heuristic_types.hpp    aliases, Achiever, FFForwardResult
  heuristic_forward.hpp  ff_forward_phase, compute_optimistic_costs
  heuristic_prob.hpp     probabilistic retry delta (now self-contained)
  heuristic_backward.hpp augment_at_with_found, ff_backward_optimistic
  heuristic.hpp          umbrella: query helpers + ff_heuristic

augment_at_with_found moves next to its sole caller. goal.hpp is
included at the top of the umbrella, removing the circular-dependency
workaround. Code bodies moved verbatim; no logic changes. Consumers
(planner.hpp, _bindings.cpp) still include heuristic.hpp unchanged.
